fmt + sort
