+++ Cargo.toml
@@ -3,6 +3,7 @@ members = [
     "account_traits_derive",
     "bytemuck",
     "cli",
+    "cli_args",
     "clients/rust/restaking_client",
     "clients/rust/vault_client",
     "core",
@@ -45,6 +46,7 @@ bytemuck = { version = "1.16.3", features = ["min_const_generics"] }
 cfg-if = "1.0.0"
 chrono = "0.4.38"
 clap = { version = "4.5.16", features = ["derive"] }
+clap-markdown = { version = "0.1.4" }
 const_str_to_pubkey = "0.1.1"
 envfile = "0.2.1"
 env_logger = "0.10.2"
@@ -52,6 +54,7 @@ jito-bytemuck = { path = "bytemuck", version = "=0.0.1" }
 jito-account-traits-derive = { path = "account_traits_derive", version = "=0.0.1" }
 jito-jsm-core = { path = "core", version = "=0.0.1" }
 jito-restaking-client = { path = "clients/rust/restaking_client", version = "=0.0.1" }
+jito-restaking-cli-args = { path = "cli_args", version = "=0.0.1" }
 jito-restaking-core = { path = "restaking_core", version = "=0.0.1" }
 jito-restaking-program = { path = "restaking_program", version = "=0.0.1" }
 jito-restaking-sdk = { path = "restaking_sdk", version = "=0.0.1" }
