+++ cli/src/bin/main.rs
@@ -1,48 +1,19 @@
-use std::{path::PathBuf, str::FromStr};
+use std::str::FromStr;
 
 use anyhow::anyhow;
 use clap::Parser;
 use env_logger::Env;
 use jito_restaking_cli::{
-    cli_args::{CliConfig, ProgramCommand},
-    restaking_handler::RestakingCliHandler,
-    vault_handler::VaultCliHandler,
+    restaking_handler::RestakingCliHandler, vault_handler::VaultCliHandler, CliConfig,
 };
+use jito_restaking_cli_args::cli_args::{Cli, ProgramCommand};
 use jito_restaking_client::programs::JITO_RESTAKING_ID;
 use jito_vault_client::programs::JITO_VAULT_ID;
 use solana_cli_config::Config;
 use solana_program::pubkey::Pubkey;
 use solana_sdk::{commitment_config::CommitmentConfig, signature::read_keypair_file};
 
-#[derive(Parser)]
-#[command(author, version, about = "A CLI for managing restaking and vault operations", long_about = None)]
-struct Cli {
-    #[command(subcommand)]
-    command: ProgramCommand,
-
-    #[arg(long, global = true, help = "Path to the configuration file")]
-    config_file: Option<PathBuf>,
-
-    #[arg(long, global = true, help = "RPC URL to use")]
-    rpc_url: Option<String>,
-
-    #[arg(long, global = true, help = "Commitment level")]
-    commitment: Option<String>,
-
-    #[arg(long, global = true, help = "Restaking program ID")]
-    restaking_program_id: Option<String>,
-
-    #[arg(long, global = true, help = "Vault program ID")]
-    vault_program_id: Option<String>,
-
-    #[arg(long, global = true, help = "Keypair")]
-    keypair: Option<String>,
-
-    #[arg(long, global = true, help = "Verbose mode")]
-    verbose: bool,
-}
-
-fn get_cli_config(args: &Cli) -> Result<CliConfig, anyhow::Error> {
+pub fn get_cli_config(args: &Cli) -> Result<CliConfig, anyhow::Error> {
     let cli_config = if let Some(config_file) = &args.config_file {
         let config = Config::load(config_file.as_os_str().to_str().unwrap())?;
         CliConfig {
@@ -94,6 +65,11 @@ async fn main() -> Result<(), anyhow::Error> {
 
     let args: Cli = Cli::parse();
 
+    if args.markdown_help {
+        clap_markdown::print_help_markdown::<Cli>();
+        return Ok(());
+    }
+
     let cli_config = get_cli_config(&args)?;
 
     let restaking_program_id = if let Some(restaking_program_id) = &args.restaking_program_id {
@@ -108,7 +84,7 @@ async fn main() -> Result<(), anyhow::Error> {
         JITO_VAULT_ID
     };
 
-    match args.command {
+    match args.command.expect("Command not found") {
         ProgramCommand::Restaking { action } => {
             RestakingCliHandler::new(cli_config, restaking_program_id, vault_program_id)
                 .handle(action)
