+++ cli/src/restaking_handler.rs
@@ -1,8 +1,10 @@
 use std::str::FromStr;
 
 use anyhow::{anyhow, Result};
-use clap::Subcommand;
 use jito_bytemuck::{AccountDeserialize, Discriminator};
+use jito_restaking_cli_args::restaking::{
+    ConfigActions, NcnActions, OperatorActions, RestakingCommands,
+};
 use jito_restaking_client::instructions::{
     InitializeConfigBuilder, InitializeNcnBuilder, InitializeOperatorBuilder,
 };
@@ -20,54 +22,7 @@ use solana_sdk::{
     transaction::Transaction,
 };
 
-use crate::cli_args::CliConfig;
-
-/// The CLI handler for the restaking program
-#[derive(Subcommand)]
-pub enum RestakingCommands {
-    /// Initialize, get, and set the config struct
-    Config {
-        #[command(subcommand)]
-        action: ConfigActions,
-    },
-    Ncn {
-        #[command(subcommand)]
-        action: NcnActions,
-    },
-    Operator {
-        #[command(subcommand)]
-        action: OperatorActions,
-    },
-}
-
-/// The actions that can be performed on the restaking config
-#[derive(Subcommand)]
-pub enum ConfigActions {
-    /// Initialize the config
-    Initialize,
-    /// Get the config
-    Get,
-}
-
-#[derive(Subcommand)]
-pub enum NcnActions {
-    /// Initialize NCN
-    Initialize,
-    /// Get NCN
-    Get { pubkey: String },
-    /// List all NCNs
-    List,
-}
-
-#[derive(Subcommand)]
-pub enum OperatorActions {
-    /// Initialize Operator
-    Initialize,
-    /// Get operator
-    Get { pubkey: String },
-    /// List all operators
-    List,
-}
+use crate::CliConfig;
 
 pub struct RestakingCliHandler {
     cli_config: CliConfig,
