+++ cli_args/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
-name = "jito-restaking-cli"
-description = "Jito Restaking CLI"
+name = "jito-restaking-cli-args"
+description = "Jito Restaking CLI arguments"
 version = { workspace = true }
 authors = { workspace = true }
 repository = { workspace = true }
@@ -15,10 +15,8 @@ borsh = { workspace = true }
 bytemuck = { workspace = true }
 chrono = { workspace = true }
 clap = { workspace = true }
-clap-markdown = { workspace = true }
 env_logger = { workspace = true }
 jito-bytemuck = { workspace = true }
-jito-restaking-cli-args = { workspace = true }
 jito-restaking-client = { workspace = true }
 jito-restaking-core = { workspace = true }
 jito-vault-client = { workspace = true }
@@ -34,10 +32,3 @@ spl-associated-token-account = { workspace = true }
 spl-token = { workspace = true }
 thiserror = { workspace = true }
 tokio = { workspace = true }
-
-[dev-dependencies]
-assert_matches = { workspace = true }
-
-[[bin]]
-name = "jito-restaking-cli"
-path = "src/bin/main.rs"
