+++ clients/js/vault_client/instructions/initializeVault.ts
@@ -93,12 +93,14 @@ export type InitializeVaultInstructionData = {
   depositFeeBps: number;
   withdrawalFeeBps: number;
   rewardFeeBps: number;
+  decimals: number;
 };
 
 export type InitializeVaultInstructionDataArgs = {
   depositFeeBps: number;
   withdrawalFeeBps: number;
   rewardFeeBps: number;
+  decimals: number;
 };
 
 export function getInitializeVaultInstructionDataEncoder(): Encoder<InitializeVaultInstructionDataArgs> {
@@ -108,6 +110,7 @@ export function getInitializeVaultInstructionDataEncoder(): Encoder<InitializeVa
       ['depositFeeBps', getU16Encoder()],
       ['withdrawalFeeBps', getU16Encoder()],
       ['rewardFeeBps', getU16Encoder()],
+      ['decimals', getU8Encoder()],
     ]),
     (value) => ({ ...value, discriminator: INITIALIZE_VAULT_DISCRIMINATOR })
   );
@@ -119,6 +122,7 @@ export function getInitializeVaultInstructionDataDecoder(): Decoder<InitializeVa
     ['depositFeeBps', getU16Decoder()],
     ['withdrawalFeeBps', getU16Decoder()],
     ['rewardFeeBps', getU16Decoder()],
+    ['decimals', getU8Decoder()],
   ]);
 }
 
@@ -153,6 +157,7 @@ export type InitializeVaultInput<
   depositFeeBps: InitializeVaultInstructionDataArgs['depositFeeBps'];
   withdrawalFeeBps: InitializeVaultInstructionDataArgs['withdrawalFeeBps'];
   rewardFeeBps: InitializeVaultInstructionDataArgs['rewardFeeBps'];
+  decimals: InitializeVaultInstructionDataArgs['decimals'];
 };
 
 export function getInitializeVaultInstruction<
